refactor(ui): extract folder-drag orchestration into a hook
