+++ docs/plan/3232-chat-artifacts-folders/README.md
@@ -51,7 +51,7 @@ All paths under `packages/ui/src/modules/artifacts/` unless noted.
 
 | #  | Title | Scope | Depends on |
 |----|-------|-------|------------|
-| 01 | [Extract folder-drag orchestration](./01-extract-drag-orchestration.md) | Refactor: lift the library view's drag state into a reusable hook; behavior unchanged | — |
+| 01 | ✅ [Extract folder-drag orchestration](./01-extract-drag-orchestration.md) | Refactor: lift the library view's drag state into a reusable hook; behavior unchanged | — |
 | 02 | [Folder groups in the chat panel](./02-panel-folder-groups.md) | Group panel rows by folder; all user folders; collapse with per-agent persisted state; counts | — |
 | 03 | [Drag-to-folder in the panel](./03-panel-drag-to-folder.md) | Rows draggable, folder headers drop targets, drop moves the artifact | 01, 02 |
 | 04 | [Marquee title truncation](./04-marquee-titles.md) | Looping title scroll on overflow per the design spec, reduced-motion safe | 02 |
